R7: Fix some macOS warnings and arm atomic
* Fix some macOS warnings wrt pthread_self
* Arm arch now needs -latomic in makefile to link
* Got test tools to successfully compile in macOS
* Fix readlink issues in macOS so now tests can run

+++ bdb/locks.c
@@ -271,7 +271,7 @@ int berkdb_lock_random_rowlock(bdb_state_type *bdb_state, int lid, int flags,
     static __thread int rowlock_rand_seq = 0;
     int *p = (int *)lkname->data;
 
-    (*p) = pthread_self();
+    (*p) = (int)pthread_self();
     p = (int *)(&((char *)lkname->data)[24]);
     (*p) = rowlock_rand_seq++;
